Failing to update questionnaire properties
They are failing because for some reason cypress is clearing out the
localStorage so the first couple of graphql calls fail - it then
succeeds.

Cypress is supposed to fully clear out localStorage between tests (`it`)
but seems to do it badly. They want to give users control over this but
that is coming later.

To resolve this issue for the moment is to cache the access token and
re-add it at the start of each test.